refactor: prepare system to additional auth provider
